暂时跑通

+++ app/src/main/java/com/example/sunflower_java/data/GardenPlantingDao.java
@@ -28,7 +28,7 @@ public interface GardenPlantingDao {
     LiveData<List<PlantAndGardenPlantings>> getPlantedGardens();
 
     @Insert
-    int insertGardenPlanting(GardenPlanting gardenPlanting);
+    long insertGardenPlanting(GardenPlanting gardenPlanting);
 
     @Delete
     void deleteGardenPlanting(GardenPlanting gardenPlanting);
